move the util source code to a directory
