+++ deep_interpolate.py
@@ -4,9 +4,9 @@ import argparse
 from tqdm import tqdm
 from interpolate_engine import InterpolateEngine
 from interpolate import Interpolate
-from simple_log import SimpleLog
-from simple_utils import max_steps
-from file_utils import create_directory
+from webui_utils.simple_log import SimpleLog
+from webui_utils.simple_utils import max_steps
+from webui_utils.file_utils import create_directory
 from typing import Callable
 
 def main():
