+++ interpolate_series.py
@@ -3,8 +3,8 @@ from tqdm import tqdm
 from interpolate_engine import InterpolateEngine
 from interpolate import Interpolate
 from deep_interpolate import DeepInterpolate
-from simple_log import SimpleLog
-from file_utils import create_directory, get_files
+from webui_utils.simple_log import SimpleLog
+from webui_utils.file_utils import create_directory, get_files
 from typing import Callable
 
 def main():
